% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webr.h.R
\name{webr}
\alias{webr}
\title{Web R}
\usage{
webr(data)
}
\arguments{
\item{data}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
}
}
\description{
Web R
}
